refactor(interop): group game adapters
