+++ CruiserJumpPractice/Interop/Game/Adapters/GameObjectAdapter.cs
@@ -7,9 +7,9 @@ using LethalCompany;
 using LethalCompany::GameNetcodeStuff;
 using LethalCompany::Unity.Netcode;
 
-using CruiserJumpPractice.Interop;
+using CruiserJumpPractice.Interop.Game;
 
-namespace CruiserJumpPractice.Interop.Adapters;
+namespace CruiserJumpPractice.Interop.Game.Adapters;
 
 internal sealed class GameObjectAdapter
 {
