+++ CruiserJumpPractice/Interop/Game/Adapters/RpcSurrogateAdapter.cs
@@ -1,11 +1,11 @@
 #nullable enable
 
 using BepInEx.Logging;
-using CruiserJumpPractice.Interop.Behaviours;
+using CruiserJumpPractice.Interop.Game.Behaviours;
 
-using CruiserJumpPractice.Interop;
+using CruiserJumpPractice.Interop.Game;
 
-namespace CruiserJumpPractice.Interop.Adapters;
+namespace CruiserJumpPractice.Interop.Game.Adapters;
 
 internal sealed class RpcSurrogateAdapter
 {
